Fix issue preventing different instances of library to work together. (closes #2)
Fixes #2

This code ensures that different instances of the library can work together when encrypting/decrypting.

Previously, when using encryption/decryption feature, an internal field (`defaultCipher`) was initialised in the encryption step. Unfortunately, this field was checked when decrypting, so if one instance set up encryption environment and a different one wanted to decrypt something, this different instance would fail due it didn't have that field initialised.

This code removes that internal instance (and therefore its check) and only checks the temporal ones created in encryption/decryption process. This ensures that different instances of the library can work together.

Thanks @MBontekoning for raising original issue.